+++ src/constants/index.ts
@@ -9,3 +9,4 @@ export * from './JIRA_TREVOR';
 export * from './JIRA_DEV1';
 export * from './JIRA_JULIE';
 export * from './JIRA_OPERATIONAL';
+export * from './JULIES_BACKGROUND';